Move addressing step library under lowering
