+++ src/lowering/ldEncodingRegMemHelpers.ts
@@ -1,4 +1,4 @@
-import type { StepPipeline } from '../addressing/steps.js';
+import type { StepPipeline } from './steps.js';
 import type { AsmOperandNode, SourceSpan } from '../frontend/ast.js';
 import type { LdForm } from './ldFormSelection.js';
 import type { LdEncodingContext } from './ldEncoding.js';
