+++ test/pr407_step_matrix.test.ts
@@ -35,7 +35,7 @@ import {
   CALC_EA,
   renderStepPipeline,
   type StepPipeline,
-} from '../src/addressing/steps';
+} from '../src/lowering/steps';
 
 const asm = renderStepPipeline;
