+++ test/pr819_exact_scale_lowering.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from 'vitest';
 
-import { EAW_GLOB_CONST, renderStepPipeline } from '../src/addressing/steps.js';
+import { EAW_GLOB_CONST, renderStepPipeline } from '../src/lowering/steps.js';
 import { DiagnosticIds, type Diagnostic } from '../src/diagnosticTypes.js';
 import type { AsmOperandNode, EaExprNode, SourceSpan, TypeExprNode } from '../src/frontend/ast.js';
 import { createAddressingPipelineBuilders } from '../src/lowering/addressingPipelines.js';
